+++ apps/assets/api/automations/gather_accounts.py
@@ -1,12 +1,12 @@
 # -*- coding: utf-8 -*-
 #
-from orgs.mixins.api import OrgBulkModelViewSet
-
-from assets.models import GatherAccountsAutomation
 from assets import serializers
+from assets.models import GatherAccountsAutomation
+from orgs.mixins.api import OrgBulkModelViewSet
+from .base import AutomationExecutionViewSet
 
 __all__ = [
-    'GatherAccountsAutomationViewSet',
+    'GatherAccountsAutomationViewSet', 'GatherAccountsExecutionViewSet'
 ]
 
 
@@ -16,3 +16,11 @@ class GatherAccountsAutomationViewSet(OrgBulkModelViewSet):
     search_fields = filter_fields
     ordering_fields = ('name',)
     serializer_class = serializers.GatherAccountAutomationSerializer
+
+
+class GatherAccountsExecutionViewSet(AutomationExecutionViewSet):
+    rbac_perms = (
+        ("list", "assets.view_gatheraccountsexecution"),
+        ("retrieve", "assets.view_gatheraccountsexecution"),
+        ("create", "assets.add_gatheraccountsexecution"),
+    )
