--- apps/assets/automations/push_account/database/mongodb/main.yml
@@ -1,16 +1,0 @@
-- hosts: mongodb
-  gather_facts: no
-  vars:
-    ansible_python_interpreter: /usr/local/bin/python
-
-  tasks:
-    - name: Add user account.username
-      mongodb_user:
-        login_user: "{{ jms_account.username }}"
-        login_password: "{{ jms_account.secret }}"
-        login_host: "{{ jms_asset.address }}"
-        login_port: "{{ jms_asset.port }}"
-        login_database: "{{ jms_asset.specific.db_name }}"
-        db: "{{ jms_asset.specific.db_name }}"
-        name: "{{ account.username }}"
-        password: "{{ account.secret }}"
