--- apps/assets/automations/push_account/database/mongodb/manifest.yml
@@ -1,6 +1,0 @@
-id: push_account_mongodb
-name: Push account from MongoDB
-category: database
-type:
-  - mongodb
-method: push_account
