--- apps/assets/automations/push_account/database/mysql/manifest.yml
@@ -1,6 +1,0 @@
-id: push_account_mysql
-name: Push account from MySQL
-category: database
-type:
-  - mysql
-method: push_account
