--- apps/assets/automations/push_account/database/oracle/manifest.yml
@@ -1,6 +1,0 @@
-id: push_account_oracle
-name: Push account from Oracle
-category: database
-type:
-  - oracle
-method: push_account
