--- apps/assets/automations/push_account/database/postgresql/manifest.yml
@@ -1,6 +1,0 @@
-id: push_account_postgresql
-name: Push account for PostgreSQL
-category: database
-type:
-  - postgresql
-method: push_account
