--- apps/assets/automations/push_account/host/posix/manifest.yml
@@ -1,7 +1,0 @@
-id: push_account_posix
-name: Push posix account
-category: host
-type:
-  - linux
-  - unix
-method: push_account
