--- apps/assets/automations/push_account/host/windows/main.yml
@@ -1,13 +1,0 @@
-- hosts: windows
-  gather_facts: yes
-  tasks:
-    - name: Add user account.username
-      ansible.windows.win_user:
-      vars:
-        fullname: "{{ account.username }}"
-        name: "{{ account.username }}"
-        password: "{{ account.secret }}"
-        state: present
-        password_expired: no
-        update_password: always
-        password_never_expires: yes
