+++ apps/assets/serializers/platform.py
@@ -45,22 +45,21 @@ class PlatformAutomationSerializer(serializers.ModelSerializer):
             "ansible_enabled", "ansible_config",
             "ping_enabled", "ping_method",
             "gather_facts_enabled", "gather_facts_method",
-            "push_account_enabled", "push_account_method",
             "change_secret_enabled", "change_secret_method",
             "verify_account_enabled", "verify_account_method",
             "gather_accounts_enabled", "gather_accounts_method",
         ]
         extra_kwargs = {
             "ping_enabled": {"label": "启用资产探测"},
-            "ping_method": {"label": "探测方式"},
-            "gather_facts_enabled": {"label": "启用收集信息"},
+            "ping_method": {"label": "资产探测方式"},
+            "gather_facts_enabled": {"label": "收集资产信息"},
             "gather_facts_method": {"label": "收集信息方式"},
             "verify_account_enabled": {"label": "启用校验账号"},
             "verify_account_method": {"label": "校验账号方式"},
             "push_account_enabled": {"label": "启用推送账号"},
             "push_account_method": {"label": "推送账号方式"},
             "change_secret_enabled": {"label": "启用账号改密"},
-            "change_secret_method": {"label": "账号创建改密方式"},
+            "change_secret_method": {"label": "账号改密方式"},
             "gather_accounts_enabled": {"label": "启用账号收集"},
             "gather_accounts_method": {"label": "收集账号方式"},
         }
